runtime(netrw): NetrwChgPerm for files not in cwd
Problem:
Changing permissions fail when using `gp` if the file under the cursor
is not in the current working directory.

Solution:
Use the already available `a:curdir` argument and prepend it to the
`<cfile>`, so that the path of the file is correct.

This commit also refactors some leftover `netrw#ErrorMsg` to
`netrw#msg#Notify` (the main refactoring was done in
f5e3b5c04f85b0f69cd2aae81e4938cfb191a790).

Signed-off-by: Mohammad Reza Karimi <m.r.karimi.j@gmail.com>