Fix building on Mac OS X 10.6 with MacPorts
Closes #17678. The original author of sound support on Mac OS (#11274) wrote
some checks that would skip the sound feature from being built if the older,
stock version of clang was used on Snow Leopard. However, if MacPorts is used
to build on 10.6 with a newer version of clang, it fails because of the use of
generics and subscripting.

This change takes this case into consideration, and allows vim +huge +sound to
build on 10.6 using MacPorts by utilizing the older style of Objective-C
syntax.

Signed-off-by: Luke Lollard <haihige@protonmail.ch>