runtime(haskell): allow spaces in backticked operators
This formatting (although rare) is actually accepted by GHC, but vim does not
highlight it. This patch adds the simplest possible regex to support the
behavior.

Inconveniently, this might trigger weird formatting on lines that contain
errors, e.g. if the first backtick is removed from:

    a `b` c `d` e

then `c` is going to be marked as an operator, which seems weird but is valid.