runtime: add popup_during_fullscreen all in i3 but not in sway config
Seems like sway won't implement `all` option for
popup_during_fullscreen, so we need to make the `all` option for
popup_during_fullscreen in sway config highlight as an error.

Reference about sway not implement `all` option for
popup_during_fullscreen:
https://github.com/swaywm/sway/issues/8746

With that in mind, we can use syntax cluster in i3config.vim so that
we can combine the shared and extra options, and then when sourcing
i3config.vim in swayconfig.vim, we can remove the extra options from
the syntax cluster.

Signed-off-by: Robertus Chris <robertusdchris@gmail.com>