do not copy writes to the '+' register to the '*' register
Before this change, writes to '+' get copied to '*', but only under
certain conditions. By default, this does not happen, because clipboard
autoselect is enabled. Disabling autoselect (an option which should only
apply to visual mode) results in normal-mode writes such as "+yy also
going to the '*' register.

This behavior is undocumented and untested; remove the behavior in order
to make Vim's handling of these two registers be consistent.

This frees up the did_star variable to be removed.

Add a test to check that the registers are independant.

fixes: #18830

Signed-off-by: Corey Hickey <bugfood-c@fatooh.org>