Fix ANSI colors in terminal on Windows
Problem: ANSI escape colors are not displayed correctly in
nontermguicolors in vim (cli) on Windows. The red and blue channels seem
to be swapped.

Cause: When converting VTerm ANSI index colors to cterm colors in
terminal.c, the Windows case equivalent to NR-16 (:help cterm-colors) is
ignored.

Solution: Created and used a table to convert ANSI indexed colors to
cterm's NR-16 representation (Windows only). This table corresponds to
the inverse conversion of cterm_ansi_idx in term.c. The values in both
tables are exactly the same, but the meanings are opposite, so they are
separate tables.

Signed-off-by: MURAOKA Taro <koron.kaoriya@gmail.com>