tests: test_sound.vim fails on Windows
Problem: tests: test_sound.vim fails when run locally on Windows

Cause: There are two causes.

First. The global variable g:id is undefined, causing an undefined
reference. Due to the execution order of test cases, g:id is defined in
Test_play_event. However, on Windows, this test is skipped, so g:id is
not defined. It is referenced in Test_play_silent's WaitForAssert
without being defined, resulting in an undefined error.

Solution: Define g:id at the beginning of Test_play_silent.

Second. In the non-GUI Windows version of vim, there is no message loop,
so the callback when play file sound ends does not occur, and
Test_play_silent's WaitForAssert times out and fails. In CI,
sound_playfile() returns 0, so Test_play_silent is skipped. The reason
for this is unknown, but it may be because CI is running on Windows
Server or something like that.

Solution: Skip Test_play_silent in Windows non-GUI environments.

Signed-off-by: MURAOKA Taro <koron.kaoriya@gmail.com>