tests: Fix flaky Test_terminal_cwd in ConPTY
Problem: tests: Test_terminal_cwd in test_terminal.vim fails flaky
in the Windows ConPTY terminal.

Processes that terminate too quickly in the ConPTY terminal cause Vim to
miss their final output.

In my environment, the probability of the "cmd /D /c cd" used in
Test_terminal_cwd occurring is about 1/4.  For a simple statically
linked Hello World, the probability of it occurring is about 3/4.

Solution: In ConPTY, the timeout is extended to 1msec when reading a
channel associated with a job that is about to finish.  This allows Vim
to read the last output of a process in a pseudo console.

In investigating this issue, I added comments to explain my
understanding of two poorly explained fields in the channel structure.

Signed-off-by: MURAOKA Taro <koron.kaoriya@gmail.com>